+++ CHANGES.rst
@@ -1,8 +1,11 @@
 Changes
 =======
 
-Unreleased
-----------
+0.3.5
+-----
+- add normal ranges
+- separate utest-id and field name
+- rename to reference_range_collection
 
 0.1.34
 ------